Change SlotRenderer connect options to use props deep equality comparison ignoring functions
